Set banner expiry date
